Rename domain_impl.rs in note_encryption to domain.rs
