+++ src/note_encryption.rs
@@ -19,7 +19,7 @@ pub mod action;
 pub mod note_encryption_vanilla;
 pub mod note_encryption_zsa;
 
-mod domain_impl;
+mod domain;
 mod note_bytes;
 mod orchard_domain;
