+++ src/note_encryption/domain.rs
